Extract GitHub pull request executor
