+++ docs/DECISIONS.md
@@ -221,3 +221,4 @@
 - Claude design critique on 2026-06-23 identified the next high-impact UI cleanup: split model selection from approval mode, replace the flat Tools popover with a structured activity surface, teach Auto/Review/Read-only as first-class controls in the empty state, and make approvals a persistent inline queue banner instead of a modal-feeling interruption. Treat those as UX roadmap inputs, not remote-executor scope.
 - Tool-step execution lives in `AgentToolStepRunner.swift`, not in `Agent.swift`. `AgentRunner.send` owns the high-level model/tool loop and repeated-call fallback; the extracted runner owns safety review, tool lifecycle event emission, unavailable-tool results, apply-patch diff follow-up execution, and tool feedback serialization.
 - Git tool schema declarations live in `GitToolDefinitions.swift`, not in `GitToolExecutor.swift`. The executor owns process execution and safety validation; the definitions catalog owns local git, GitHub PR, and worktree tool names, descriptions, JSON schemas, host, and risk metadata.
+- GitHub pull request execution lives in `GitHubPullRequestToolExecutor.swift`, and raw git/gh process launching lives in `GitProcessRunner.swift`. `GitToolExecutor` remains the compatibility facade used by the tool router and remote planners, but PR command construction, PR-specific input validation, URL artifact extraction, and raw `Process` management should stay out of that facade.
